refactor: 카카오 맵 API 호출 시, 위도&경도 값 추가 저장(#229)

+++ src/main/java/org/scoula/dto/PlaceDTO.java
@@ -9,8 +9,8 @@ public class PlaceDTO {
     private String placeType;
     private String placeName;
     private String address;
-    private Integer distance;           // 거리 (meter)
-    private double latitude;
-    private double longitude;
+    private Long distance;
+    private Double latitude;
+    private Double longitude;
     private String categoryName;
 }
